PPDS assgt3/requirements.txt
requests
pandas
beautifulsoup4
lxml
python-dotenv
tenacity
urllib3<2.0,>=1.26
